refactor(rename files and packages) (#20)
