+++ src/main.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { Provider } from 'react-redux';
-import { store } from './app/store.js';
+import { store } from './redux/store.js';
 import App from './App';
 import './index.css';
