+++ src/redux/actions/transactionActions.js
@@ -1,5 +1,5 @@
 import { createAsyncThunk } from '@reduxjs/toolkit';
-import { apiPrivate } from '../../service/getData';
+import { apiPrivate } from '../../service/httpRequest';
 
 export const getTransactions = createAsyncThunk('transactions/getTransactions', async () => {
   try {
